+++ Runtime/CyberiadaGraph/EdgeData.cs
@@ -1,7 +1,8 @@
 using System.Collections.Generic;
+using System.Text;
 using UnityEngine;
 
-namespace Talent.Graph.Cyberiada
+namespace Talent.Graphs
 {
     public class EdgeData : IClonable<EdgeData>
     {
@@ -77,6 +78,22 @@ namespace Talent.Graph.Cyberiada
         }
 
         #endregion
+
+        public override string ToString()
+        {
+            StringBuilder stringBuilder = new();
+            stringBuilder.AppendLine($"EdgeData({TriggerID})({Condition})({VisualData.Position})");
+
+            string actions = "\n";
+            foreach (Action action in Actions)
+            {
+                actions += $"\n{action}";
+            }
+
+            stringBuilder.AppendLine(actions);
+
+            return stringBuilder.ToString();
+        }
     }
 
     public class EdgeVisualData