+++ Runtime/EditorCore/GraphEditor.cs
@@ -1,29 +1,28 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
-using Talent.Graph;
-using Talent.Graph.Cyberiada;
-using Action = Talent.Graph.Cyberiada.Action;
-using Event = Talent.Graph.Cyberiada.Event;
+using Talent.Graphs;
+using Action = Talent.Graphs.Action;
 
 namespace Talent.GraphEditor.Core
 {
     public class GraphEditor
     {
-        public Graph<GraphData, NodeData, EdgeData> Graph { get; private set; }
+        public CyberiadaGraph Graph { get; private set; }
 
         private IGraphElementViewFactory GraphElementViewFactory { get; }
-        private Dictionary<string, Node<GraphData, NodeData, EdgeData>> _nodes = new Dictionary<string, Node<GraphData, NodeData, EdgeData>>();
-        private Dictionary<string, Edge<EdgeData>> _edges = new Dictionary<string, Edge<EdgeData>>();
+        private Dictionary<string, Node> _nodes = new Dictionary<string, Node>();
+        private Dictionary<string, Edge> _edges = new Dictionary<string, Edge>();
 
-        private BidirectionalDictionary<Graph<GraphData, NodeData, EdgeData>, IGraphView> _graphViews = new BidirectionalDictionary<Graph<GraphData, NodeData, EdgeData>, IGraphView>();
-        private BidirectionalDictionary<Node<GraphData, NodeData, EdgeData>, INodeView> _nodeViews = new BidirectionalDictionary<Node<GraphData, NodeData, EdgeData>, INodeView>();
-        private BidirectionalDictionary<Edge<EdgeData>, IEdgeView> _edgeViews = new BidirectionalDictionary<Edge<EdgeData>, IEdgeView>();
+        private BidirectionalDictionary<CyberiadaGraph, IGraphView> _graphViews = new BidirectionalDictionary<CyberiadaGraph, IGraphView>();
+        private BidirectionalDictionary<Node, INodeView> _nodeViews = new BidirectionalDictionary<Node, INodeView>();
+        private BidirectionalDictionary<Edge, IEdgeView> _edgeViews = new BidirectionalDictionary<Edge, IEdgeView>();
         private BidirectionalDictionary<Event, INodeEventView> _nodeEventViews = new BidirectionalDictionary<Event, INodeEventView>();
         private BidirectionalDictionary<Action, INodeActionView> _nodeActionViews = new BidirectionalDictionary<Action, INodeActionView>();
         private BidirectionalDictionary<Action, IEdgeActionView> _edgeActionViews = new BidirectionalDictionary<Action, IEdgeActionView>();
 
-        private Node<GraphData, NodeData, EdgeData> _initialNode;
-        private Edge<EdgeData> _initialEdge;
+        private INodeView _initialNodeView;
+        private IEdgeView _initialEdgeView;
 
         public GraphEditor(IGraphElementViewFactory factory)
         {
@@ -34,7 +33,7 @@ namespace Talent.GraphEditor.Core
         /// Set graph and rebuild views
         /// </summary>
         /// <param name="graph"></param>
-        public void SetGraph(Graph<GraphData, NodeData, EdgeData> graph)
+        public void SetGraph(CyberiadaGraph graph)
         {
             Graph = graph;
             RebuildView();
@@ -96,43 +95,23 @@ namespace Talent.GraphEditor.Core
             _nodes.Clear();
             _edges.Clear();
 
-            _initialNode = null;
-            _initialEdge = null;
+            _initialNodeView = null;
+            _initialEdgeView = null;
 
             if (Graph == null)
             {
                 return;
             }
 
-            foreach (Node<GraphData, NodeData, EdgeData> node in Graph.Nodes)
+            foreach (Node node in Graph.Nodes)
             {
                 if (node.Data.Vertex == NodeData.Vertex_Initial)
                 {
-                    _initialNode = node;
-                    _nodeViews.Set(_initialNode, GraphElementViewFactory.CreateNodeView(_initialNode.Data.VisualData, _initialNode.Data.Vertex, false));
+                    _initialNodeView = CreateViewForNode(node, false);
                 }
                 else
                 {
-                    CreateViewForNode(node, false, false);
-                }
-            }
-
-            foreach (Edge<EdgeData> edge in Graph.Edges)
-            {
-                if (edge.SourceNode == _initialNode.ID || edge.TargetNode == _initialNode.ID)
-                {
-                    _initialEdge = edge;
-
-                    if (_nodeViews.TryGetValue(_initialEdge.SourceNode == _initialNode.ID ? _initialNode : _nodes[_initialEdge.SourceNode], out INodeView sourceNodeView)
-                        && _nodeViews.TryGetValue(_initialEdge.TargetNode == _initialNode.ID ? _initialNode : _nodes[_initialEdge.TargetNode], out INodeView targetNodeView))
-                    {
-                        IEdgeView edgeView = GraphElementViewFactory.CreateEdgeView(sourceNodeView, targetNodeView, _initialEdge.Data.VisualData, _initialEdge.Data.TriggerID, _initialEdge.Data.Condition);
-                        _edgeViews.Add(_initialEdge, edgeView);
-                    }
-                }
-                else
-                {
-                    CreateViewForEdge(edge);
+                    CreateViewForNode(node, false);
                 }
             }
 
@@ -140,34 +119,83 @@ namespace Talent.GraphEditor.Core
             {
                 CreateInitialNode();
             }
+
+            foreach (Edge edge in Graph.Edges)
+            {
+                if (edge.SourceNode == NodeData.Vertex_Initial)
+                {
+                    _initialEdgeView = CreateViewForEdge(edge);
+                }
+                else
+                {
+                    CreateViewForEdge(edge);
+                }
+            }
+        }
+
+        public void ConnectToInitialNode(INodeView nodeView)
+        {
+            if (_initialNodeView == null)
+            {
+                return;
+            }
+
+            if (_initialEdgeView != null)
+            {
+                RemoveEdge(_initialEdgeView);
+            }
+
+            _initialEdgeView = CreateNewEdge(_initialNodeView, nodeView, "");
+        }
+
+        public bool TryDuplicateNode(INodeView nodeView, out INodeView duplicatedNode)
+        {
+            if (!_nodeViews.TryGetValue(nodeView, out Node node))
+            {
+                duplicatedNode = null;
+                return false;
+            }
+
+            Node newNode = node.GetCopy(node.Data.GetCopy(), parentNode: node.ParentNode, newID: Guid.NewGuid().ToString());
+
+            Graph.AddNode(newNode);
+
+            duplicatedNode = CreateViewForNode(newNode, true);
+            return true;
         }
 
         private void CreateInitialNode()
         {
-            if (_initialNode == null)
+            if (_initialNodeView == null)
             {
-                _initialNode = new Node<GraphData, NodeData, EdgeData>("initial", new NodeData(NodeData.Vertex_Initial));
+                Node nodeData = new(NodeData.Vertex_Initial, new NodeData(NodeData.Vertex_Initial));
 
-                Graph.AddNode(_initialNode);
-                _initialNode.Data.VisualData.Name = "INIT";
+                Graph.AddNode(nodeData);
+                nodeData.Data.VisualData.Name = "INIT";
 
-                _nodeViews.Set(_initialNode, GraphElementViewFactory.CreateNodeView(_initialNode.Data.VisualData, _initialNode.Data.Vertex, true));
-
-                CreateInitialEdge();
+                _initialNodeView = CreateViewForNode(nodeData, true);
+                _nodeViews.Set(nodeData, _initialNodeView);
             }
         }
 
-        private void CreateInitialEdge()
+        private void CreateRandomInitialEdge()
         {
-            if (_initialEdge == null && _initialNode != null && _nodes.Count > 0)
+            if (_initialEdgeView == null && _initialNodeView != null)
             {
-                _initialEdge = new Edge<EdgeData>(_initialNode.ID + _nodes.Values.First().ID, _initialNode.ID, _nodes.Values.First().ID, new EdgeData(""));
-                Graph.AddEdge(_initialEdge);
+                Node targetNode = _nodes.Values.Where((Node node) => node.Data.Vertex != NodeData.Vertex_Initial).FirstOrDefault();
 
-                if (_nodeViews.TryGetValue(_initialNode, out INodeView sourceNodeView) && _nodeViews.TryGetValue(_nodes.Values.First(), out INodeView targetNodeView))
+                if (targetNode == null)
                 {
-                    IEdgeView edgeView = GraphElementViewFactory.CreateEdgeView(sourceNodeView, targetNodeView, _initialEdge.Data.VisualData, _initialEdge.Data.TriggerID, _initialEdge.Data.Condition);
-                    _edgeViews.Add(_initialEdge, edgeView);
+                    return;
+                }
+
+                Edge edge = new Edge(NodeData.Vertex_Initial + targetNode.ID, NodeData.Vertex_Initial, targetNode.ID, new EdgeData(""));
+                Graph.AddEdge(edge);
+
+                if (_nodeViews.TryGetValue(targetNode, out INodeView targetNodeView))
+                {
+                    _initialEdgeView = GraphElementViewFactory.CreateEdgeView(_initialNodeView, targetNodeView, edge.Data.VisualData, edge.Data.TriggerID, edge.Data.Condition);
+                    _edgeViews.Add(edge, _initialEdgeView);
                 }
             }
         }
@@ -176,9 +204,9 @@ namespace Talent.GraphEditor.Core
 
         public IGraphView CreateNewGraph(string id, INodeView nodeView)
         {
-            if (_nodeViews.TryGetValue(nodeView, out Node<GraphData, NodeData, EdgeData> node) && node.NestedGraph == null)
+            if (_nodeViews.TryGetValue(nodeView, out Node node) && node.NestedGraph == null)
             {
-                node.NestedGraph = new Graph<GraphData, NodeData, EdgeData>(id, new GraphData());
+                node.NestedGraph = new CyberiadaGraph(id, new GraphData());
             }
 
             IGraphView graphView = GraphElementViewFactory.CreateGraphView(id, nodeView);
@@ -188,24 +216,32 @@ namespace Talent.GraphEditor.Core
 
         public INodeView CreateNewNode(string name)
         {
-            Node<GraphData, NodeData, EdgeData> node = new Node<GraphData, NodeData, EdgeData>(System.Guid.NewGuid().ToString(), new NodeData());
+            Node node = new Node(System.Guid.NewGuid().ToString(), new NodeData());
             Graph.AddNode(node);
             node.Data.VisualData.Name = name;
 
-            return CreateViewForNode(node, true, true);
+            INodeView newNode = CreateViewForNode(node, true);
+
+            if (_initialNodeView == null)
+            {
+                CreateInitialNode();
+                ConnectToInitialNode(newNode);
+            }
+
+            return newNode;
         }
 
         public IEdgeView CreateNewEdge(INodeView sourceView, INodeView targetView, string triggerID)
         {
-            Node<GraphData, NodeData, EdgeData> sourceNode = _nodeViews.Get(sourceView);
-            Node<GraphData, NodeData, EdgeData> targetNode = _nodeViews.Get(targetView);
+            Node sourceNode = _nodeViews.Get(sourceView);
+            Node targetNode = _nodeViews.Get(targetView);
 
-            if ((sourceNode == _initialNode || targetNode == _initialNode) && _initialEdge != null)
+            if ((sourceNode == _initialNodeView || targetNode == _initialNodeView) && _initialEdgeView != null)
             {
                 return null;
             }
 
-            Edge<EdgeData> edge = new Edge<EdgeData>(System.Guid.NewGuid().ToString(), sourceNode.ID, targetNode.ID, new EdgeData(triggerID));
+            Edge edge = new Edge(System.Guid.NewGuid().ToString(), sourceNode.ID, targetNode.ID, new EdgeData(triggerID));
             Graph.AddEdge(edge);
 
             return CreateViewForEdge(edge);
@@ -215,17 +251,14 @@ namespace Talent.GraphEditor.Core
         {
             INodeEventView result = default;
 
-            if (_nodeViews.TryGetValue(nodeView, out Node<GraphData, NodeData, EdgeData> node))
+            if (_nodeViews.TryGetValue(nodeView, out Node node))
             {
-                if (!node.Data.Events.TryGetValue(triggerID, out Event nodeEvent))
-                {
-                    nodeEvent = new Event(triggerID);
-                    node.Data.AddEvent(nodeEvent);
-                }
+                Event nodeEvent = new Event(triggerID);
+                node.Data.AddEvent(nodeEvent);
 
                 if (!_nodeEventViews.ContainsKey(nodeEvent))
                 {
-                    _nodeEventViews.Set(nodeEvent, GraphElementViewFactory.CreateNodeEventView(_nodeViews.Get(node), triggerID));
+                    _nodeEventViews.Set(nodeEvent, GraphElementViewFactory.CreateNodeEventView(_nodeViews.Get(node), triggerID, nodeEvent));
                 }
 
                 result = _nodeEventViews.Get(nodeEvent);
@@ -254,7 +287,7 @@ namespace Talent.GraphEditor.Core
         {
             IEdgeActionView result = default;
 
-            if (_edgeViews.TryGetValue(edgeView, out Edge<EdgeData> edge))
+            if (_edgeViews.TryGetValue(edgeView, out Edge edge))
             {
                 Action eventAction = new Action(actionID);
 
@@ -268,28 +301,15 @@ namespace Talent.GraphEditor.Core
 
         #endregion
 
-        public void ChangeNodeEventTrigger(INodeView nodeView, INodeEventView nodeEventView, string triggerID)
+        public void ChangeNodeEventTrigger(INodeEventView nodeEventView, Event @event, string triggerID)
         {
-            if (_nodeViews.TryGetValue(nodeView, out Node<GraphData, NodeData, EdgeData> node) && _nodeEventViews.TryGetValue(nodeEventView, out Event nodeEvent) && nodeEvent.TriggerID != triggerID)
-            {
-                if (!node.Data.Events.TryGetValue(triggerID, out Event newNodeEvent) || !_nodeEventViews.TryGetValue(newNodeEvent, out INodeEventView newNodeEventView))
-                {
-                    newNodeEventView = CreateNewNodeEvent(nodeView, triggerID);
-                }
-
-                foreach (Action action in nodeEvent.Actions)
-                {
-                    var actionView = CreateNewNodeAction(newNodeEventView, action.ID);
-                    ChangeNodeActionParameter(actionView, action.Parameter);
-                }
-
-                RemoveNodeEvent(nodeView, nodeEventView);
-            }
+            @event.SetTrigger(triggerID);
+            nodeEventView.SetTrigger(triggerID);
         }
 
         public void ChangeEdgeTrigger(IEdgeView edgeView, string triggerID)
         {
-            if (_edgeViews.TryGetValue(edgeView, out Edge<EdgeData> edge))
+            if (_edgeViews.TryGetValue(edgeView, out Edge edge))
             {
                 edge.Data.SetTrigger(triggerID);
                 edgeView.SetTrigger(triggerID);
@@ -298,7 +318,7 @@ namespace Talent.GraphEditor.Core
 
         public void ChangeNodeEventCondition(INodeView nodeView, INodeEventView nodeEventView, string condition)
         {
-            if (_nodeViews.TryGetValue(nodeView, out Node<GraphData, NodeData, EdgeData> node) && _nodeEventViews.TryGetValue(nodeEventView, out Event nodeEvent) && nodeEvent.Condition != condition)
+            if (_nodeViews.TryGetValue(nodeView, out Node node) && _nodeEventViews.TryGetValue(nodeEventView, out Event nodeEvent) && nodeEvent.Condition != condition)
             {
                 nodeEvent.SetCondition(condition);
                 nodeEventView.SetCondition(condition);
@@ -307,28 +327,28 @@ namespace Talent.GraphEditor.Core
 
         public void ChangeEdgeCondition(IEdgeView edgeView, string condition)
         {
-            if (_edgeViews.TryGetValue(edgeView, out Edge<EdgeData> edge))
+            if (_edgeViews.TryGetValue(edgeView, out Edge edge))
             {
                 edge.Data.SetCondition(condition);
                 edgeView.SetCondition(condition);
             }
         }
 
-        public void ChangeNodeActionParameter(INodeActionView actionView, string parameter)
+        public void ChangeNodeActionParameter(INodeActionView actionView, List<Tuple<string, string>> parameters)
         {
             if (_nodeActionViews.TryGetValue(actionView, out Action action))
             {
-                action.SetParameter(parameter);
-                actionView.SetParameter(parameter);
+                action.SetParameters(parameters);
+                actionView.SetParameters(parameters);
             }
         }
 
-        public void ChangeEdgeActionParameter(IEdgeActionView actionView, string parameter)
+        public void ChangeEdgeActionParameter(IEdgeActionView actionView, List<Tuple<string, string>> parameters)
         {
             if (_edgeActionViews.TryGetValue(actionView, out Action action))
             {
-                action.SetParameter(parameter);
-                actionView.SetParameter(parameter);
+                action.SetParameters(parameters);
+                actionView.SetParameters(parameters);
             }
         }
 
@@ -336,13 +356,13 @@ namespace Talent.GraphEditor.Core
 
         public void RemoveNode(INodeView nodeView, bool createInitialEdge = true)
         {
-            if (_nodeViews.TryGetValue(nodeView, out Node<GraphData, NodeData, EdgeData> node))
+            if (_nodeViews.TryGetValue(nodeView, out Node node))
             {
-                Node<GraphData, NodeData, EdgeData> oldParent = node.ParentNode;
+                Node oldParent = node.ParentNode;
 
-                foreach (KeyValuePair<string, Event> nodeEvent in node.Data.Events)
+                foreach (Event nodeEvent in node.Data.Events)
                 {
-                    foreach (Action action in nodeEvent.Value.Actions)
+                    foreach (Action action in nodeEvent.Actions)
                     {
                         INodeActionView nodeActionView = _nodeActionViews.Get(action);
 
@@ -350,14 +370,14 @@ namespace Talent.GraphEditor.Core
                         GraphElementViewFactory.DestroyElementView(nodeActionView);
                     }
 
-                    if (_nodeEventViews.TryGetValue(nodeEvent.Value, out INodeEventView eventView))
+                    if (_nodeEventViews.TryGetValue(nodeEvent, out INodeEventView eventView))
                     {
-                        _nodeEventViews.Remove(nodeEvent.Value);
+                        _nodeEventViews.Remove(nodeEvent);
                         GraphElementViewFactory.DestroyElementView(eventView);
                     }
                 }
 
-                foreach (Edge<EdgeData> edge in Graph.Edges.Where(edge => edge.SourceNode == node.ID || edge.TargetNode == node.ID).ToArray())
+                foreach (Edge edge in Graph.Edges.Where(edge => edge.SourceNode == node.ID || edge.TargetNode == node.ID).ToArray())
                 {
                     if (_edgeViews.TryGetValue(edge, out IEdgeView edgeView))
                     {
@@ -365,7 +385,7 @@ namespace Talent.GraphEditor.Core
                         {
                             IEdgeActionView edgeActionView = _edgeActionViews.Get(action);
 
-                            if (_edgeViews.TryGetValue(edgeView, out Edge<EdgeData> nodeEvent) && _edgeActionViews.TryGetValue(edgeActionView, out Action eventAction))
+                            if (_edgeViews.TryGetValue(edgeView, out Edge nodeEvent) && _edgeActionViews.TryGetValue(edgeActionView, out Action eventAction))
                             {
                                 _edgeActionViews.Remove(edgeActionView);
                                 GraphElementViewFactory.DestroyElementView(edgeActionView);
@@ -377,16 +397,16 @@ namespace Talent.GraphEditor.Core
                         Graph.DeleteEdge(edge);
                         GraphElementViewFactory.DestroyElementView(edgeView);
 
-                        if (edge.SourceNode == _initialNode.ID || edge.TargetNode == _initialNode.ID)
+                        if (edge.SourceNode == NodeData.Vertex_Initial || edge.TargetNode == NodeData.Vertex_Initial)
                         {
-                            _initialEdge = null;
+                            _initialEdgeView = null;
                         }
                     }
                 }
 
                 if (node.NestedGraph != null)
                 {
-                    foreach (Node<GraphData, NodeData, EdgeData> child in node.NestedGraph.Nodes.ToArray())
+                    foreach (Node child in node.NestedGraph.Nodes.ToArray())
                     {
                         if (_nodeViews.TryGetValue(child, out INodeView childView))
                         {
@@ -405,19 +425,19 @@ namespace Talent.GraphEditor.Core
                 _nodeViews.Remove(node);
                 GraphElementViewFactory.DestroyElementView(nodeView);
 
-                if (_nodes.Count > 0)
+                if (_nodes.Count > 1)
                 {
                     if (createInitialEdge)
                     {
-                        CreateInitialEdge();
+                        CreateRandomInitialEdge();
                     }
                 }
                 else
                 {
-                    Graph.DeleteNode(_initialNode);
-                    GraphElementViewFactory.DestroyElementView(_nodeViews.Get(_initialNode));
-                    _nodeViews.Remove(_initialNode);
-                    _initialNode = null;
+                    Graph.DeleteNode(_nodes[NodeData.Vertex_Initial]);
+                    GraphElementViewFactory.DestroyElementView(_initialNodeView);
+                    _nodeViews.Remove(_initialNodeView);
+                    _initialNodeView = null;
                 }
 
                 RemoveGraphFromNode(oldParent);
@@ -426,7 +446,7 @@ namespace Talent.GraphEditor.Core
 
         public void RemoveEdge(IEdgeView edgeView)
         {
-            if (_edgeViews.TryGetValue(edgeView, out Edge<EdgeData> edge))
+            if (_edgeViews.TryGetValue(edgeView, out Edge edge))
             {
                 foreach (Action action in edge.Data.Actions)
                 {
@@ -445,7 +465,7 @@ namespace Talent.GraphEditor.Core
 
         public void RemoveNodeEvent(INodeView nodeView, INodeEventView nodeEventView)
         {
-            Node<GraphData, NodeData, EdgeData> node = _nodeViews.Get(nodeView);
+            Node node = _nodeViews.Get(nodeView);
             Event nodeEvent = _nodeEventViews.Get(nodeEventView);
 
             foreach (Action action in nodeEvent.Actions)
@@ -478,7 +498,7 @@ namespace Talent.GraphEditor.Core
 
         public void RemoveEdgeAction(IEdgeView edgeView, IEdgeActionView edgeActionView)
         {
-            if (_edgeViews.TryGetValue(edgeView, out Edge<EdgeData> nodeEvent) && _edgeActionViews.TryGetValue(edgeActionView, out Action eventAction))
+            if (_edgeViews.TryGetValue(edgeView, out Edge nodeEvent) && _edgeActionViews.TryGetValue(edgeActionView, out Action eventAction))
             {
                 nodeEvent.Data.RemoveAction(eventAction);
                 _edgeActionViews.Remove(edgeActionView);
@@ -490,12 +510,12 @@ namespace Talent.GraphEditor.Core
 
         public void SetParent(INodeView childView, INodeView parentView, bool layoutAutomatically)
         {
-            if (childView == null || !_nodeViews.TryGetValue(childView, out Node<GraphData, NodeData, EdgeData> child) || (_initialNode != null && child == _initialNode))
+            if (childView == null || !_nodeViews.TryGetValue(childView, out Node child) || (_initialNodeView != null && child == _initialNodeView))
             {
                 return;
             }
 
-            Node<GraphData, NodeData, EdgeData> oldParent = null;
+            Node oldParent = null;
 
             if (parentView == null)
             {
@@ -510,7 +530,7 @@ namespace Talent.GraphEditor.Core
 
                 _nodeViews.Get(child).SetParent(null, layoutAutomatically);
             }
-            else if (_nodeViews.TryGetValue(parentView, out Node<GraphData, NodeData, EdgeData> parent) && child != parent && (_initialNode == null || parent != _initialNode))
+            else if (_nodeViews.TryGetValue(parentView, out Node parent) && child != parent && (_initialNodeView == null || parent != _initialNodeView))
             {
                 if (child.ParentNode != parent)
                 {
@@ -549,7 +569,8 @@ namespace Talent.GraphEditor.Core
             RemoveGraphFromNode(oldParent);
         }
 
-        private bool NodeHasNodeRecursively(Node<GraphData, NodeData, EdgeData> node1, Node<GraphData, NodeData, EdgeData> node2)
+        // TODO: remove recursion
+        private bool NodeHasNodeRecursively(Node node1, Node node2)
         {
             if (node1 == null || node1.NestedGraph == null || node2 == null)
             {
@@ -561,7 +582,7 @@ namespace Talent.GraphEditor.Core
                 return true;
             }
 
-            foreach (Node<GraphData, NodeData, EdgeData> nestedNode in node1.NestedGraph.Nodes)
+            foreach (Node nestedNode in node1.NestedGraph.Nodes)
             {
                 if (NodeHasNodeRecursively(nestedNode, node2))
                 {
@@ -572,7 +593,7 @@ namespace Talent.GraphEditor.Core
             return false;
         }
 
-        private void RemoveGraphFromNode(Node<GraphData, NodeData, EdgeData> node)
+        private void RemoveGraphFromNode(Node node)
         {
             if (node == null || node.NestedGraph == null)
             {
@@ -593,29 +614,30 @@ namespace Talent.GraphEditor.Core
 
         #region Internal Create Methods
 
-        private INodeView CreateViewForNode(Node<GraphData, NodeData, EdgeData> node, bool createInitialNode, bool layoutAutomatically)
+        private INodeView CreateViewForNode(Node node, bool layoutAutomatically)
         {
             _nodes[node.ID] = node;
             INodeView view = GraphElementViewFactory.CreateNodeView(node.Data.VisualData, node.Data.Vertex, layoutAutomatically);
             _nodeViews.Set(node, view);
 
-            foreach (KeyValuePair<string, Event> nodeEvent in node.Data.Events)
+            foreach (Event nodeEvent in node.Data.Events)
             {
-                string[] actionsArray = new string[nodeEvent.Value.Actions.Count];
+                string[] actionsArray = new string[nodeEvent.Actions.Count];
 
                 for (int i = 0; i < actionsArray.Length; i++)
                 {
-                    actionsArray[i] = nodeEvent.Value.Actions[i].ID;
+                    actionsArray[i] = nodeEvent.Actions[i].ID;
                 }
 
-                INodeEventView eventView = GraphElementViewFactory.CreateNodeEventView(view, nodeEvent.Key);
-                _nodeEventViews.Add(nodeEvent.Value, eventView);
+                INodeEventView eventView = GraphElementViewFactory.CreateNodeEventView(view, nodeEvent.TriggerID, nodeEvent);
+                _nodeEventViews.Add(nodeEvent, eventView);
+                eventView.SetCondition(nodeEvent.Condition);
 
-                foreach (Action action in nodeEvent.Value.Actions)
+                foreach (Action action in nodeEvent.Actions)
                 {
                     var actionView = GraphElementViewFactory.CreateNodeActionView(eventView, action.ID);
                     _nodeActionViews.Add(action, actionView);
-                    ChangeNodeActionParameter(actionView, action.Parameter);
+                    ChangeNodeActionParameter(actionView, action.Parameters);
                 }
             }
 
@@ -623,25 +645,17 @@ namespace Talent.GraphEditor.Core
             {
                 CreateNewGraph(node.NestedGraph.ID, view);
 
-                foreach (Node<GraphData, NodeData, EdgeData> nestedNode in node.NestedGraph.Nodes)
+                foreach (Node nestedNode in node.NestedGraph.Nodes)
                 {
-                    INodeView nestedView = CreateViewForNode(nestedNode, false, layoutAutomatically);
+                    INodeView nestedView = CreateViewForNode(nestedNode, layoutAutomatically);
                     SetParent(nestedView, view, layoutAutomatically);
-                }
-            }
-
-            if (createInitialNode)
-            {
-                if (_initialNode == null && _nodes.Count > 0)
-                {
-                    CreateInitialNode();
                 }
             }
 
             return view;
         }
 
-        private IEdgeView CreateViewForEdge(Edge<EdgeData> edge)
+        private IEdgeView CreateViewForEdge(Edge edge)
         {
             IEdgeView edgeView = null;
 
@@ -655,7 +669,7 @@ namespace Talent.GraphEditor.Core
                 {
                     var actionView = GraphElementViewFactory.CreateEdgeActionView(edgeView, action.ID);
                     _edgeActionViews.Add(action, actionView);
-                    ChangeEdgeActionParameter(actionView, action.Parameter);
+                    ChangeEdgeActionParameter(actionView, action.Parameters);
                 }
             }
 
@@ -671,7 +685,7 @@ namespace Talent.GraphEditor.Core
     {
         IGraphView CreateGraphView(string graphID, INodeView parentNodeView);
         INodeView CreateNodeView(NodeVisualData visualData, string vertex, bool layoutAutomatically);
-        INodeEventView CreateNodeEventView(INodeView nodeView, string triggerID);
+        INodeEventView CreateNodeEventView(INodeView nodeView, string triggerID, Event @event);
         INodeActionView CreateNodeActionView(INodeEventView eventView, string actionID);
         IEdgeView CreateEdgeView(INodeView sourceNode, INodeView targetNode, EdgeVisualData edgeVisualData, string triggerID, string condition);
         IEdgeActionView CreateEdgeActionView(IEdgeView edgeView, string actionID);
@@ -690,11 +704,12 @@ namespace Talent.GraphEditor.Core
     public interface INodeEventView : IGraphElementView
     {
         void SetCondition(string condition);
+        void SetTrigger(string triggerID);
     }
 
     public interface INodeActionView : IGraphElementView
     {
-        void SetParameter(string parameter);
+        void SetParameters(List<Tuple<string, string>> parameters);
     }
 
     public interface IEdgeView : IGraphElementView
@@ -705,7 +720,7 @@ namespace Talent.GraphEditor.Core
 
     public interface IEdgeActionView : IGraphElementView
     {
-        void SetParameter(string parameter);
+        void SetParameters(List<Tuple<string, string>> parameters);
     }
 
     #endregion